check all combinations of format bits for BCH code
There are two copies of the format bits, so iterate through all 2^15
options of {some of the bits from the first copy, some of the bits from
the second copy} by brute force.

I haven't implemented BCH *error correction* yet, so this will not be
able to correct the case where *the same* format bit is flipped in both
copies.  But so far, this seems to be correctly fixing a bit along the X
that was drawn over the QR code in the original photo.

However, this seems to have also started causing the Rust side to panic
when the JS calls update().